Add files via upload
This is a basic personal portfolio website built using HTML and CSS. It includes a homepage with a short introduction, sections for projects, contact information, and responsive styling for different screen sizes. It's designed to be clean, beginner-friendly, and easy to customize. I created this to showcase my journey in web development and keep track of my progress as I continue learning.